pyproject.toml
[project]
name = "openpi-client"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "dm-tree>=0.1.8",
    "msgpack>=1.0.5",
    "numpy>=1.22.4,<2.0.0",
    "pillow>=9.0.0",
    "tree>=0.2.4",
    "typing-extensions>=4.12.2",
    "websockets>=11.0",
    "opencv-python",
    "pyrealsense2",
    "xarm-python-sdk",
    "ipykernel",
    "polars",
    "rich",
    "tqdm",
    "tyro",
]

[dependency-groups]
dev = ["pytest>=8.3.4"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py310"
